# libs/@local/graph/api/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "graph",
    "description": "HASH Graph API",
    "contact": {
      "name": "HASH"
    },
    "license": {
      "name": "AGPL-3"
    },
    "version": "0.0.0"
  },
  "paths": {
    "/account_groups": {
      "post": {
        "tags": [
          "Graph",
          "Account Group"
        ],
        "operationId": "create_account_group",
        "parameters": [
          {
            "name": "X-Authenticated-User-Actor-Id",
            "in": "header",
            "description": "The ID of the actor which is used to authorize the request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UntaggedActorId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InsertAccountGroupIdParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The schema of the created account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UntaggedTeamId"
                }
              }
            }
          },
          "500": {
            "description": "Store error occurred"
          }
        }
      }
    },
    "/account_groups/{account_group_id}/members/{account_id}": {
      "post": {
        "tags": [
          "Graph",
          "Account Group"
        ],
        "operationId": "add_account_group_member",
        "parameters": [
          {
            "name": "X-Authenticated-User-Actor-Id",
            "in": "header",
            "description": "The ID of the actor which is used to authorize the request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UntaggedActorId"
            }
          },
          {
            "name": "account_group_id",
            "in": "path",
            "description": "The ID of the account group to add the member to",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UntaggedTeamId"
            }
          },
          {
            "name": "account_id",
            "in": "path",
            "description": "The ID of the account to add to the group",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UntaggedActorId"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "The account group member was added"
          },
          "403": {
            "description": "Permission denied"
          },
          "500": {
            "description": "Store error occurred"
          }
        }
      },
      "delete": {
        "tags": [
          "Graph",
          "Account Group"
        ],
        "operationId": "remove_account_group_member",
        "parameters": [
          {
            "name": "X-Authenticated-User-Actor-Id",
            "in": "header",
            "description": "The ID of the actor which is used to authorize the request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UntaggedActorId"
            }
          },
          {
            "name": "account_group_id",
            "in": "path",
            "description": "The ID of the account group to remove the member from",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UntaggedTeamId"
            }
          },
          {
            "name": "account_id",
            "in": "path",
            "description": "The ID of the account to remove from the group",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UntaggedActorId"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The account group member was removed"
          },
          "403": {
            "description": "Permission denied"
          },
          "500": {
            "description": "Store error occurred"
          }
        }
      }
    },
    "/account_groups/{account_group_id}/permissions/{permission}": {
      "get": {
        "tags": [
          "Graph",
          "Account Group"
        ],
        "operationId": "check_account_group_permission",
        "parameters": [
          {
            "name": "X-Authenticated-User-Actor-Id",
            "in": "header",
            "description": "The ID of the actor which is used to authorize the request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UntaggedActorId"
            }
          },
          {
            "name": "account_group_id",
            "in": "path",
            "description": "The ID of the account group to check if the actor has the permission",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UntaggedTeamId"
            }
          },
          {
            "name": "permission",
            "in": "path",
            "description": "The permission to check for",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AccountGroupPermission"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Information if the actor can add an owner",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error occurred"
          }
        }
      }
    },
    "/account_groups/{account_group_id}/relations": {
      "get": {
        "tags": [
          "Graph",
          "Account Group"
        ],
        "operationId": "get_account_group_relations",
        "parameters": [
          {
            "name": "X-Authenticated-User-Actor-Id",
            "in": "header",
            "description": "The ID of the actor which is used to authorize the request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UntaggedActorId"
            }
          },
          {
            "name": "account_group_id",
            "in": "path",
            "description": "The ID of the account group to get relations from",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UntaggedTeamId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of members and administrators of the account group",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccountGroupRelationAndSubject"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Permission denied"
          },
          "500": {
            "description": "Internal error occurred"
          }
        }
      }
    },
    "/accounts": {
      "post": {
        "tags": [
          "Graph",
          "Account"
        ],
        "operationId": "create_account",
        "parameters": [
          {
            "name": "X-Authenticated-User-Actor-Id",
            "in": "header",
            "description": "The ID of the actor which is used to authorize the request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UntaggedActorId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InsertAccountIdParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The schema of the created account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UntaggedActorId"
                }
              }
            }
          },
          "500": {
            "description": "Store error occurred"
          }
        }
      }
    },
...
    "/entities/query/count": {
      "post": {
        "tags": [
          "Graph",
          "Entity"
        ],
        "operationId": "count_entities",
        "parameters": [
          {
            "name": "X-Authenticated-User-Actor-Id",
            "in": "header",
            "description": "The ID of the actor which is used to authorize the request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UntaggedActorId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CountEntitiesParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "minimum": 0
                }
              }
            }
          },
          "422": {
            "description": "Provided query is invalid"
          },
          "500": {
            "description": "Store error occurred"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountGroupAdministratorSubject": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "subjectId",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "account"
                ]
              },
              "subjectId": {
                "$ref": "#/components/schemas/UntaggedActorId"
              }
            }
          }
        ],
        "discriminator": {
          "propertyName": "kind"
        }
      },
      "AccountGroupMemberSubject": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "subjectId",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "account"
                ]
              },
              "subjectId": {
                "$ref": "#/components/schemas/UntaggedActorId"
              }
            }
          }
        ],
        "discriminator": {
          "propertyName": "kind"
        }
      },
      "AccountGroupPermission": {
        "type": "string",
        "enum": [
          "add_member",
          "remove_member"
        ]
      },
      "AccountGroupRelationAndSubject": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "subject",
              "relation"
            ],
            "properties": {
              "relation": {
                "type": "string",
                "enum": [
                  "administrator"
                ]
              },
              "subject": {
                "$ref": "#/components/schemas/AccountGroupAdministratorSubject"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "subject",
              "relation"
            ],
            "properties": {
              "relation": {
                "type": "string",
                "enum": [
                  "member"
                ]
              },
              "subject": {
                "$ref": "#/components/schemas/AccountGroupMemberSubject"
              }
            }
          }
        ],
        "discriminator": {
          "propertyName": "relation"
        }
      },
...
      "CountEntitiesParams": {
        "type": "object",
        "required": [
          "filter",
          "temporalAxes",
          "includeDrafts"
        ],
        "properties": {
          "filter": {
            "$ref": "#/components/schemas/Filter"
          },
          "includeDrafts": {
            "type": "boolean"
          },
          "temporalAxes": {
            "$ref": "#/components/schemas/QueryTemporalAxesUnresolved"
          }
        },
        "additionalProperties": false
      },
...
      "Filter": {
        "oneOf": [
          {
            "type": "object",
            "title": "AllFilter",
            "required": [
              "all"
            ],
            "properties": {
              "all": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Filter"
                }
              }
            }
          },
          {
            "type": "object",
            "title": "AnyFilter",
            "required": [
              "any"
            ],
            "properties": {
              "any": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Filter"
                }
              }
            }
          },
          {
            "type": "object",
            "title": "NotFilter",
            "required": [
              "not"
            ],
            "properties": {
              "not": {
                "$ref": "#/components/schemas/Filter"
              }
            }
          },
          {
            "type": "object",
            "title": "EqualFilter",
            "required": [
              "equal"
            ],
            "properties": {
              "equal": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/FilterExpression"
                },
                "maxItems": 2,
                "minItems": 2
              }
            }
          },
          {
            "type": "object",
            "title": "NotEqualFilter",
            "required": [
              "notEqual"
            ],
            "properties": {
              "notEqual": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/FilterExpression"
                },
                "maxItems": 2,
                "minItems": 2
              }
            }
          },
          {
            "type": "object",
            "title": "GreaterFilter",
            "required": [
              "notEqual"
            ],
            "properties": {
              "greater": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/FilterExpression"
                },
                "maxItems": 2,
                "minItems": 2
              }
            }
          },
          {
            "type": "object",
            "title": "GreaterOrEqualFilter",
            "required": [
              "notEqual"
            ],
            "properties": {
              "greaterOrEqual": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/FilterExpression"
                },
                "maxItems": 2,
                "minItems": 2
              }
            }
          },
          {
            "type": "object",
            "title": "LessFilter",
            "required": [
              "notEqual"
            ],
            "properties": {
              "less": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/FilterExpression"
                },
                "maxItems": 2,
                "minItems": 2
              }
            }
          },
          {
            "type": "object",
            "title": "LessOrEqualFilter",
            "required": [
              "notEqual"
            ],
            "properties": {
              "lessOrEqual": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/FilterExpression"
                },
                "maxItems": 2,
                "minItems": 2
              }
            }
          },
          {
            "type": "object",
            "title": "CosineDistanceFilter",
            "required": [
              "cosineDistance"
            ],
            "properties": {
              "cosineDistance": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/FilterExpression"
                },
                "maxItems": 3,
                "minItems": 3
              }
            }
          },
          {
            "type": "object",
            "title": "StartsWithFilter",
            "required": [
              "startsWith"
            ],
            "properties": {
              "startsWith": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/FilterExpression"
                },
                "maxItems": 2,
                "minItems": 2
              }
            }
          },
          {
            "type": "object",
            "title": "EndsWithFilter",
            "required": [
              "endsWith"
            ],
            "properties": {
              "endsWith": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/FilterExpression"
                },
                "maxItems": 2,
                "minItems": 2
              }
            }
          },
          {
            "type": "object",
            "title": "ContainsSegmentFilter",
            "required": [
              "containsSegment"
            ],
            "properties": {
              "containsSegment": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/FilterExpression"
                },
                "maxItems": 2,
                "minItems": 2
              }
            }
          }
        ]
      },
...
      "InsertAccountGroupIdParams": {
        "type": "object",
        "properties": {
          "accountGroupId": {
            "$ref": "#/components/schemas/UntaggedTeamId"
          }
        },
        "additionalProperties": false
      },
      "InsertAccountIdParams": {
        "type": "object",
        "properties": {
          "accountId": {
            "$ref": "#/components/schemas/UntaggedActorId"
          }
        },
        "additionalProperties": false
      },
...
      "PermissionResponse": {
        "type": "object",
        "required": [
          "has_permission"
        ],
        "properties": {
          "has_permission": {
            "type": "boolean"
          }
        }
      },
...
      "UntaggedActorId": {
        "type": "string",
        "format": "uuid"
      },
      "UntaggedTeamId": {
        "type": "string",
        "format": "uuid"
      },
...
    }
  },
...
}